src/utils/email_alert.py
import smtplib
import os
from email.mime.text import MIMEText
from dotenv import load_dotenv

load_dotenv()

def send_email_alert(message):
    msg = MIMEText(message)
    msg['Subject'] = os.getenv("EMAIL_SUBJECT")
    msg['From'] = os.getenv("EMAIL_FROM")
    msg['To'] = os.getenv("EMAIL_TO")

    try:
        with smtplib.SMTP(os.getenv("SMTP_SERVER"), int(os.getenv("SMTP_PORT"))) as server:
            server.starttls()
            server.login(os.getenv("EMAIL_USER"), os.getenv("EMAIL_PASS"))
            server.send_message(msg)
            print("📧 Alert email sent.")
    except Exception as e:
        print(f"❌ Failed to send email: {e}")
